Handel the ctrl+c of the application
The select options are little TEA apps. They were capturing the ctrl+c
and returning values, so you could not quit out of the app.

I changed it so that ctrl+c returns "true" for quit and then just quit
the app from there.

There might be a better way.